inlined memory read/write functions and modifications to cpu::emulate() based on perf report
inlined memory functions since they get called so frequently.
changed op in cpu::emulate() to int since it's used to address an array later on.
changed dt from uint8 to int to prevent having to zero-extend registers for no reason.
changed cycles/cb_cycles to int since they are used to increment dt.
changed length and interrupt_table to uint16 since they are used to increment pc.w

trying to implement a cpu throttling and experimentally looking for ideal
settings for minimum CPU usage. for reference mednafen idles at about 15% when
running joyp.gb. extensively tested usleep but it seems that more granular
control would be better so I'm switching to nanosleep/clock_nanosleep.
this is very much a work in progress

+++ cpu.h
@@ -16,9 +16,9 @@ class cpu
     void print_registers(void);
 };
 
-const static int interrupt_table[16] = {0,0x08,0,0x10,0,0,0,0x18,0,0,0,0,0,0,0,0x20};
+const static uint16 interrupt_table[16] = {0,0x08,0,0x10,0,0,0,0x18,0,0,0,0,0,0,0,0x20};
 
-const static uint8 length[0x0100] = 
+const static uint16 length[0x0100] = 
 {
 /*0 1 2 3 4 5 6 7 8 9 A B C D E F*/
   1,3,1,1,1,1,2,1,3,1,1,1,1,1,2,1, /*0x*/
@@ -39,9 +39,9 @@ const static uint8 length[0x0100] =
   2,1,1,1,0,1,2,1,2,1,3,1,0,0,2,1  /*Fx*/
 };
 
-const static uint8 cb_cycles[0x08] = {2,2,2,2,2,2,4,2};
+const static int cb_cycles[0x08] = {2,2,2,2,2,2,4,2};
 
-const static uint8 cycles[0x0100] =
+const static int cycles[0x0100] =
 {
 /*0 1 2 3 4 5 6 7 8 9 A B C D E F*/
   1,3,2,2,1,1,2,1,5,2,2,2,1,1,2,1, /*0x*/
